requirements.txt
cached-property==1.5.2
certifi==2020.11.8
chardet==3.0.4
Django==2.2.13
django-filter==2.4.0
django-url-filter==0.3.15
djangorestframework==3.9.4
enum-compat==0.0.3
httpie==2.3.0
idna==2.10
Pygments==2.7.2
PySocks==1.7.1
pytz==2020.4
requests==2.25.0
requests-toolbelt==0.9.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
